move util functions to util.py
